Fixed degradation
We needed to add an offset to the 'store64' address to aquire an address inside the loop. Degrade can now take an optional offset argument (default 25) and run_degradation.sh takes a list of taskset masks to run degrade on (i.e. run it as ./run_degradation.sh 0x8 0x80 0x1 0x2)